requirements.txt
fastapi==0.115.12
starlette==0.46.2
uvicorn==0.34.3
sqlalchemy==2.0.41
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.16.1
passlib==1.7.4
bcrypt==4.3.0
python-jose==3.5.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
aiohttp==3.12.13
aiosmtplib==4.0.1
itsdangerous==2.2.0
python-dotenv==1.1.0
httpx==0.28.1
requests==2.32.4
edge-tts==7.0.2
googletrans==4.0.2
faster-whisper==1.0.1
jose==1.0.0
typing_extensions==4.14.0
greenlet==3.2.3
multidict==6.4.4
yarl==1.20.1
frozenlist==1.7.0
sniffio==1.3.1
anyio==4.9.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
hyperframe==6.1.0
websockets==15.0.1
aiosignal==1.3.2
aiohappyeyeballs==2.6.1
colorama==0.4.6
certifi==2025.6.15
cffi==1.17.1
pycparser==2.22
six==1.17.0
setuptools==80.9.0
attrs==25.3.0
pydantic[email]
tiktoken
redis>=4.2.0
langdetect
